Handle new format of diagnostic reporting.
On the latest version of Ghostty on Mac (as of 27-01-2025) the format for reporting diagnostic seems to be different than what the plugin expects. Therefore, it fails to parse the diagnostic output, and fails to present any diagnostics in neovim. A simple update of the supported regex fixes the issue.


> nvim --version                                                                                                                             NVIM v0.10.3
Build type: Release
LuaJIT 2.1.1736781742
Run "nvim -V1 -v" for more info

> ghostty --version                                                                                                                          Ghostty 1.0.1

Version
  - version: 1.0.1
  - channel: stable
Build Config
  - Zig version: 0.13.0
  - build mode : builtin.OptimizeMode.ReleaseFast
  - app runtime: apprt.Runtime.none
  - font engine: font.main.Backend.coretext
  - renderer   : renderer.Metal
  - libxev     : main.Backend.kqueue